Streamline code.
