+++ src/main/java/org/moqucu/games/nightstalker/event/CollisionListener.java
@@ -1,8 +1,8 @@
 package org.moqucu.games.nightstalker.event;
 
-import org.moqucu.games.nightstalker.model.GameObjectImpl;
+import org.moqucu.games.nightstalker.model.GameObject;
 
 public interface CollisionListener {
 
-    void collideWith(GameObjectImpl gameObject);
+    void collideWith(GameObject gameObject);
 }
