+++ src/main/java/org/moqucu/games/nightstalker/model/GameWorld.java
@@ -11,7 +11,7 @@ public class GameWorld {
     private long time = 0;
 
     @Getter
-    private final Map<String, GameObjectImpl> objects = new HashMap<>();
+    private final Map<String, GameObject> objects = new HashMap<>();
 
     @Getter
     private final Set<TimeListener> timeListeners = new HashSet<>();
@@ -25,7 +25,7 @@ public class GameWorld {
         timeListeners.forEach(timeListener -> timeListener.elapseTime(milliseconds));
     }
 
-    public void add(GameObjectImpl gameObject) {
+    public void add(GameObject gameObject) {
 
         objects.put(gameObject.getObjectId(), gameObject);
         if (gameObject instanceof TimeListener)
